refactor: IconPropType moved to search page type files and update of import statement in appropiate files
Relates #256
