+++ components/icons/BoyIcon.tsx
@@ -1,4 +1,4 @@
-import { IconPropType } from '../search/filter/QuickBrowse';
+import { IconPropType } from '@/types/searchPageTypes';
 
 const BoyIcon: React.FC<IconPropType> = ({ subcategory }) => {
   return (
