+++ components/icons/ShoeIcon.tsx
@@ -1,4 +1,4 @@
-import { IconPropType } from '../search/filter/QuickBrowse';
+import { IconPropType } from '@/types/searchPageTypes';
 
 const ShoeIcon: React.FC<IconPropType> = ({ category }) => {
   return (
