+++ types/searchPageTypes.ts
@@ -5,3 +5,11 @@ export type SearchParamsType = {
   limit: number;
   cursor: string;
 };
+
+export type IconPropType = {
+  width?: number;
+  height?: number;
+  active?: boolean;
+  category?: string;
+  subcategory?: string;
+};
